Add a regression test
