+++ tests/test-no-std/lib.rs
@@ -92,11 +92,5 @@ mod tests {
         assert_eq!(num, 7)
     }
 
-    #[test]
-    fn empty_literal() {
-        let input = "";
-        parse!(input, "");
-        let input = "Test";
-        parse!(input, "Test")
-    }
+    include!("../common.rs");
 }
